backend/requirements.txt
flask                 # Web framework
psycopg2-binary       # PostgreSQL driver
redis                 # Redis client